requirements.txt
fastapi>=0.129.0
uvicorn>=0.41.0
psychrolib>=2.5.0
pydantic>=2.12.0
scipy>=1.17.0
python-multipart>=0.0.9
fpdf2>=2.7.0
numpy>=1.26.0
pytest>=9.0.0
httpx>=0.28.0
